Dev/2025215369/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
aiosqlite==0.20.0
httpx==0.27.2
dnspython==2.6.1
ping3==4.0.4
